moved controllers to a separate file Refactored code
