+++ index.js
@@ -2,13 +2,9 @@ import express from 'express';
 import dotenv from 'dotenv';
 dotenv.config();
 import mongoose from 'mongoose';
-import UserModel from './models/User.js'
 import checkAuth from './utils/checkAuth.js'
-import jwt from 'jsonwebtoken'
-import bcrypt from 'bcrypt'
 import { registerValidation } from './validations/auth.js';
-import { validationResult } from 'express-validator';
-import User from './models/User.js';
+import { register, login, getMe } from './controllers/UserController.js';
 const app = express();
 const port = process.env.PORT;
 app.use(express.json());
@@ -17,105 +13,11 @@ mongoose
     .then(() => console.log('DB OK'))
     .catch((err) => console.log('DB ERROR', err))
 
-app.get('/', (req, res) => {
-    res.send('Working');
-})
+app.post('/auth/register', registerValidation, register);
 
-app.post('/auth/register', registerValidation, async (req, res) => {
-    try {
-        const errors = validationResult(req);
-        if(!errors.isEmpty()) {
-            return res.status(400).json(errors.array());
-        }
+app.post('/auth/login', login);
 
-        const password = req.body.password;
-        const salt = await bcrypt.genSalt(10);
-        const hash = await bcrypt.hash(password, salt);
-        const doc = new UserModel({
-            fullName: req.body.fullName,
-            email: req.body.email,
-            passwordHash: hash,
-        });
-
-        const user = await doc.save();
-        const token = jwt.sign(
-            {
-                _id: user._id,
-            },
-            process.env.ID_KEY,
-            {
-                expiresIn: '30d'
-            }
-        );
-
-        const { passwordHash, ...userData } = user._doc;
-        res.json({
-            ...userData,
-            token
-        });
-    }
-    catch (error) {
-        console.log(error);
-        res.status(500).json({
-            message: 'Не вдалося зареєструватися',
-        })
-    }
-});
-
-app.post('/auth/login', async (req, res) => {
-    try {
-        const user = await UserModel.findOne({ email: req.body.email });
-        if (!user) {
-            return res.status(400).json({
-                message: 'Невірний логін або пароль'
-            });
-        }
-        const isValidPass = await bcrypt.compare(req.body.password, user._doc.passwordHash);
-        if (!isValidPass) {
-            return res.status(400).json({
-                message: 'Невірний логін або пароль'
-            });
-        }
-        const token = jwt.sign(
-            {
-                _id: user._id,
-            },
-            process.env.ID_KEY,
-            {
-                expiresIn: '30d'
-            }
-        );
-
-        const { passwordHash, ...userData } = user._doc;
-        res.json({
-            ...userData,
-            token
-        });
-    } catch (error) {
-        console.log(error);
-        res.status(500).json({
-            message: 'Не вдалося авторизуватися',
-        })
-    }
-});
-
-app.get('/auth/me', checkAuth, async (req, res) => {
-    try {
-        const user = await UserModel.findById(req.userId);
-        if (!user) {
-            return res.status(404).json({
-                message: 'Невірний логін або пароль',
-            })
-        }
-        const { passwordHash, ...userData } = user._doc;
-        res.json(userData);
-    } catch (error) {
-        console.log(error);
-        res.status(500).json({
-            message: 'Нема доступа',
-        })
-    }
-});
+app.get('/auth/me', checkAuth, getMe);
 
 app.listen(port, (err) => {
     if (err) {
